refactor: move TitleBar to atoms
